+++ src/components/elements/Input/Title/titleInputStyle.ts
@@ -1,0 +1,18 @@
+import { css } from "@emotion/react";
+
+
+export const titleInputStyle = css`
+  width: 300px;
+  min-width: 20px;
+  margin: 1px;
+  padding: 0.2em 1em;
+  border-radius: 8px;
+  border: 1px solid transparent;
+  font-size: 1em;
+  font-weight: 200;
+  font-family: inherit;
+  color: #5b5b5b;
+  background-color: #ffffff;
+  transition: border-color 0.25s;
+  box-shadow: 0 2px 2px rgba(0, 0, 0, 0.2);
+`;
